conflict

See merge request nova/misc/wx-public/proxy!2

+++ internal/domain/entity/wx.go
@@ -18,7 +18,7 @@ type TextRequestBody struct {
 	CreateTime   int64
 	MsgType      string
 	Content      string
-	MsgID        int
+	MsgID        int64
 	Event        string
 	Ticket       string
 	EventKey     string
